requirements.txt
configparser==5.3.0
DateTime==4.7
mysql-connector==2.2.9
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
zope.interface==5.5.1
